added update and patch methods in where required

+++ src/main/java/com/nomudev/models/OrderLineModel.java
@@ -34,7 +34,7 @@ public class OrderLineModel {
     private Long orderLineId;
 
     @Column(nullable = false)
-    private int orderLineAmount;
+    private float orderLineAmount;
 
     @ManyToOne
     @JoinColumn(name = "orderId", nullable = false)
